Moved the k8s logic inside guid_pool.go to the daemon.go
Moved the InitPool() function which read the existing pods guids
and store them from the guid_pool.go to the main.go. Also moved
the guidPodNetworkMap which contains a map between the pods and
their guids from the guid_pool.go to daemon.go and made the necessary
changes.

fixes #44