Change ib-kubernetes update strategy to Recreate
When RollingUpdate strategy is used, the old pod is only deleted,
when the new pod gets to the running state to reduce the downtime.
https://kubernetes.io/docs/tutorials/kubernetes-basics/update/update-intro/
However, for ib-kubernetes we specified the anti affinity,
which prevent two ib-kubernetes pods running on the same node.
In this case, when the deployment is updated, the new pod is
forever pending because the old one is not deleted.

To fix this, changing the update strategy to Recreate.

As part of this change, drop the PodAntiAffinity as it doesn't
bring any useful impact on the deployment.

Signed-off-by: amaslennikov <amaslennikov@nvidia.com>